Feat. Typescript 변경에 따른 수정
