+++ package.json
@@ -39,7 +39,12 @@
         "html-webpack-plugin": "^5.5.3",
         "prettier": "3.2.5",
         "style-loader": "^4.0.0",
+        "ts-loader": "^9.5.1",
+        "ts-node": "^10.9.2",
         "webpack-cli": "^5.1.4",
         "webpack-dev-server": "^4.15.1"
+    },
+    "dependencies": {
+        "typescript": "^5.4.5"
     }
 }
