+++ src/apis/getSearchedMovies.js
@@ -1,11 +1,12 @@
-import ApiClient from './apiClient';
+import ApiClient from '../domain/ApiClient';
 
-async function getSearchedMovies(query) {
+async function getSearchedMovies(query, page) {
     const res = ApiClient.request({
         method: 'GET',
         endpoint: '3/search/movie',
         params: {
             query: query,
+            page: page,
             api_key: process.env.TMDB_API_KEY,
             language: 'ko-KR'
         }
